requirements.txt
llama-parse>=0.4.0
python-dotenv
